Initial rework with cookies added and URL removed (for now)
