+++ src/Abstracts/Input.php
@@ -1,9 +1,19 @@
 <?php
 
-namespace DjinnDev\RequestHandler;
+namespace DjinnDev\RequestHandler\Abstracts;
 
-abstract class AbstractInput extends AbstractBase
+use \DjinnDev\RequestHandler\Interfaces\Factory;
+
+abstract class Input implements Factory
 {
+	/**
+	 * Protected.
+	 * Used for saving static instace of class
+	 * 
+	 * @var object|null
+	 */
+	protected static ?object $_instance = null;
+	
 	/**
 	 * Protected.
 	 * An array of all the information sent over.
@@ -20,20 +30,6 @@ abstract class AbstractInput extends AbstractBase
 	 * @var array
 	 */
 	private array $__validKeyTypes = ['string' => 1, 'integer' => 1];
-	
-	/**
-	 * Protected.
-	 * Function returns trimmed content of php://input.
-	 * 
-	 * @return string
-	 */
-	protected function _getPhpInputContents(): string
-	{
-		// pull data from input
-		$input = file_get_contents('php://input');
-		// basic sanity
-		return trim($input);
-	}
 
 	/**
 	 * Public.
@@ -45,7 +41,7 @@ abstract class AbstractInput extends AbstractBase
 	{
 		return $this->_rawData;
 	}
-
+    
 	/**
 	 * Public.
 	 * Returns value of specific location in request.
